[Silver IV] Title: 격자 막기, Time: 0 ms, Memory: 3244 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/33754. 격자 막기/격자 막기.cc
@@ -2,7 +2,7 @@
 #pragma GCC optimize("O3,unroll-loops")
 #pragma GCC target("avx,avx2,fma")
 using namespace std;
-using ll = long long;
+using ll = int;
 using pll = pair<ll,ll>;
 using ld = long double;
 using pld = pair<ld,ld>;
@@ -193,10 +193,10 @@ void solve(ll testcase){
             cnt += up+down;
             continue;
         }
-        io << 0LL;
+        io << 0;
         return;
     }
-    io << (cnt==2*N ? 2LL : 1LL);
+    io << (cnt==2*N ? 2 : 1);
 }
 
 int main() {
